Redid turret statuses to be tied to the turret itself not the projectile.
By reworking where the statuses a turret has to be attached to the turret itself as opposed to its projectile, upgrading the statuses itself is now possible instead of creating a new status when you want to modify its properties. This makes upgrades easier and getting the data to show to player should now be easier as well though not yet implemented. Also added the poison status.